Update CVector page even though we don't currently use it
Currently we use the gh-pages branch of the github repo which contains a
rendered README.md as index.html, the test results and doxygen output.

I also keep that stuff here as cvector.html and in projects/cvector/
because I want to keep my site self contained so I can move to
self-hosting relatively easily if I want/need to.